requirements.txt
boto3==1.16.54
Flask==1.1.1
line-bot-sdk==1.15.0
oauth2client==4.1.3
gspread==3.6.0
PyDrive==1.3.1
imgurpython==1.1.7
